+++ tests/testthat/test-list_deps.R
@@ -1,22 +1,45 @@
 library(fs)
 
-test_that("list_deps works", {
+test_that("non parsing list_deps works", {
 
   f <- new_factory(path = path_temp(), move_in = FALSE)
   on.exit(dir_delete(f))
-  
+
   # copy test reports over (as this has inline code)
-  file_copy(
-    path("test_reports", "package_calls.Rmd"),
-    path(f, "report_sources")
-  )
-  
-  expected_deps_package_calls <- c("purrr", "readxl", "fs")
-  expected_deps_example_report <- c("here", "incidence2")
-  expected_deps_readme <- "rmarkdown"
-  expected_deps <- c(expected_deps_package_calls,
-                     expected_deps_example_report,
-                     expected_deps_readme)
+  file_copy(path("test_reports", "package_calls.Rmd"), path(f, "report_sources"))
+  file_copy(path("test_reports", "example.R"), path(f, "report_sources"))
+
+  # non parsed expected_deps
+  expected_deps_package_calls_rmd <- c("purrr", "readxl", "fs", "rmarkdown")
+  expected_deps_example_r <- c("fs", "yaml", "callr", "utils", "rprojroot")
+  expected_deps_example_report <- c("base", "knitr", "fs")
+  expected_deps <- c(expected_deps_package_calls_rmd,
+                     expected_deps_example_r,
+                     expected_deps_example_report)
+
   deps <- list_deps(f)
-  expect_equal(sort(deps), sort(expected_deps))
+  expect_equal(sort(deps), sort(unique(expected_deps)))
 })
+
+
+test_that("parsing list_deps works", {
+
+  f <- new_factory(path = path_temp(), move_in = FALSE)
+  on.exit(dir_delete(f))
+
+  # copy test reports over (as this has inline code)
+  file_copy(path("test_reports", "package_calls.Rmd"), path(f, "report_sources"))
+  file_copy(path("test_reports", "example.R"), path(f, "report_sources"))
+
+  # non parsed expected_deps
+  expected_deps_package_calls_rmd <- c("purrr", "readxl", "fs", "rmarkdown")
+  expected_deps_example_r <- c("fs", "yaml", "callr", "utils", "rprojroot")
+  expected_deps_example_report <- c("knitr", "fs")
+  expected_deps <- c(expected_deps_package_calls_rmd,
+                     expected_deps_example_r,
+                     expected_deps_example_report)
+
+  deps <- list_deps(f, parse_first = TRUE)
+  expect_equal(sort(deps), sort(unique(expected_deps)))
+})
+